requirements.txt
#pyyaml == 5.1.2
django-gravatar2 == 1.4.2
psycopg2 == 2.8.4
param == 1.9.3
django == 2.2.6
django-select2 == 7.1.2
django-colorfield == 0.3.2